Trial Balance Added

+++ src/main/java/edu/kennesaw/appdomain/service/AdminService.java
@@ -1,5 +1,6 @@
 package edu.kennesaw.appdomain.service;
 
+import edu.kennesaw.appdomain.dto.NewUserDTO;
 import edu.kennesaw.appdomain.entity.UserDate;
 import edu.kennesaw.appdomain.entity.UserSecurity;
 import edu.kennesaw.appdomain.repository.*;
@@ -9,12 +10,13 @@ import edu.kennesaw.appdomain.dto.UserDTO;
 import edu.kennesaw.appdomain.entity.User;
 import edu.kennesaw.appdomain.exception.UserAttributeMissingException;
 import edu.kennesaw.appdomain.service.utils.ServiceUtils;
+import jakarta.persistence.EntityManager;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.token.TokenService;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.io.File;
 import java.io.IOException;
@@ -66,7 +68,7 @@ public class AdminService {
     private ScriptService scriptService;
 
     @Autowired
-    private EmailService emailService;
+    private EntityManager em;
 
     public ResponseEntity<?> createUserWithRole(UserDTO userDTO) throws UserAttributeMissingException, IOException, InterruptedException {
 
@@ -94,7 +96,7 @@ public class AdminService {
 
         UserSecurity userSecurity = new UserSecurity();
         String uuid = UUID.randomUUID().toString();
-        userSecurity.setPassword(passwordEncoder.encode(uuid));
+        userSecurity.setPassword(uuid, passwordEncoder);
         userSecurity.setIsActive(true);
         userSecurity.setIsVerified(true);
         userSecurity.setFailedLoginAttempts(0);
@@ -114,13 +116,10 @@ public class AdminService {
         return ResponseEntity.ok(createdUser);
     }
 
-    public ResponseEntity<?> updateUser(UserDTO userDetails) throws IOException, InterruptedException {
+    @Transactional
+    public ResponseEntity<?> updateUser(NewUserDTO userDetails) throws IOException, InterruptedException {
 
-        if (userDetails.getUserid().isEmpty()) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new MessageResponse("Error passing userid."));
-        }
-
-        Optional<User> optionalExistingUser = userRepository.findByUserid(userDetails.getUserid().get());
+        Optional<User> optionalExistingUser = userRepository.findByUserid(userDetails.getUserid());
 
         if (optionalExistingUser.isEmpty()) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new MessageResponse("User not found."));
@@ -130,33 +129,38 @@ public class AdminService {
 
         String oldUsername = existingUser.getUsername();
 
-        if (userDetails.getUserType().isPresent()) {
-            UserType existingUserType = existingUser.getUserType();
-            UserType newUserType = userDetails.getUserType().get();
-            if (existingUserType.equals(UserType.DEFAULT) || existingUserType.equals(UserType.USER)) {
-                if (!newUserType.equals(UserType.DEFAULT) && !newUserType.equals(UserType.USER)) {
-                    scriptService.createMailbox(existingUser);
-                }
-            } else {
-                if (newUserType.equals(UserType.DEFAULT) || newUserType.equals(UserType.USER)) {
-                    scriptService.deleteMailbox(existingUser);
-                }
+        UserType existingUserType = existingUser.getUserType();
+        UserType newUserType = userDetails.getUserType();
+
+        if (existingUserType.equals(UserType.DEFAULT) || existingUserType.equals(UserType.USER)) {
+            if (!newUserType.equals(UserType.DEFAULT) && !newUserType.equals(UserType.USER)) {
+                scriptService.createMailbox(existingUser);
+            }
+        } else {
+            if (newUserType.equals(UserType.DEFAULT) || newUserType.equals(UserType.USER)) {
+                if (scriptService.mailboxExists(existingUser)) scriptService.deleteMailbox(existingUser);
             }
         }
 
-        existingUser.setEmail(userDetails.getEmail().isEmpty() ? existingUser.getEmail() : userDetails.getEmail().get());
-        existingUser.setUsername(userDetails.getUsername().isEmpty() ? existingUser.getUsername() : userDetails.getUsername().get());
-        existingUser.setFirstName(userDetails.getFirstName().isEmpty() ? existingUser.getLastName() : userDetails.getFirstName().get());
-        existingUser.setLastName(userDetails.getLastName().isEmpty() ? existingUser.getLastName() : userDetails.getLastName().get());
-        existingUser.getUserDate().setBirthday(userDetails.getBirthday().isEmpty() ? existingUser.getUserDate().getBirthday() : userDetails.getBirthday().get());
-        existingUser.getUserSecurity().setIsVerified(userDetails.getIsVerified().isEmpty() ? existingUser.getUserSecurity().getIsVerified() : userDetails.getIsVerified().get());
-        existingUser.getUserSecurity().setIsActive(userDetails.getIsActive().isEmpty() ? existingUser.getUserSecurity().getIsActive() : userDetails.getIsActive().get());
-        existingUser.setUserType(userDetails.getUserType().isEmpty() ? existingUser.getUserType() : userDetails.getUserType().get());
-        existingUser.getUserSecurity().setFailedLoginAttempts(userDetails.getFailedPasswordAttempts().isEmpty() ? existingUser.getUserSecurity().getFailedLoginAttempts() : userDetails.getFailedPasswordAttempts().get());
-        existingUser.getUserDate().setTempLeaveStart(userDetails.getTempLeaveStart().isEmpty() ? existingUser.getUserDate().getTempLeaveStart() : userDetails.getTempLeaveStart().get());
-        existingUser.getUserDate().setTempLeaveEnd(userDetails.getTempLeaveEnd().isEmpty() ? existingUser.getUserDate().getTempLeaveEnd() : userDetails.getTempLeaveEnd().get());
-        existingUser.getUserSecurity().setEmailPassword(userDetails.getEmailPassword().isEmpty() ? existingUser.getUserSecurity().getEmailPassword() : userDetails.getEmailPassword().get());
-        userRepository.save(existingUser);
+        existingUser.setEmail(userDetails.getEmail());
+        existingUser.setUsername(userDetails.getUsername());
+        existingUser.setFirstName(userDetails.getFirstName());
+        existingUser.setLastName(userDetails.getLastName());
+        existingUser.setUserType(userDetails.getUserType());
+
+        UserSecurity userSecurity = userSecurityRepository.findById(existingUser.getUserSecurity().getId()).orElse(new UserSecurity());
+        userSecurity.setIsVerified(userDetails.getUserSecurity().getIsVerified());
+        userSecurity.setIsActive(userDetails.getUserSecurity().getIsActive());
+        userSecurity.setFailedLoginAttempts(userDetails.getUserSecurity().getFailedLoginAttempts());
+
+        existingUser.setUserSecurity(userSecurityRepository.save(userSecurity));
+
+        UserDate userDate = userDateRepository.findById(existingUser.getUserDate().getId()).orElse(new UserDate());
+        userDate.setTempLeaveStart(userDetails.getUserDate().getTempLeaveStart());
+        userDate.setTempLeaveEnd(userDetails.getUserDate().getTempLeaveEnd());
+        userDate.setBirthday(userDetails.getUserDate().getBirthday());
+
+        existingUser.setUserDate(userDateRepository.save(userDate));
 
         if (!oldUsername.equals(existingUser.getUsername())) {
             scriptService.updateUsername(oldUsername.toLowerCase(), existingUser.getUsername().toLowerCase());
@@ -169,7 +173,7 @@ public class AdminService {
             Files.move(oldFilePath, newFilePath);
         }
 
-        return ResponseEntity.ok(existingUser);
+        return ResponseEntity.ok(userRepository.save(existingUser));
 
     }
 
@@ -181,63 +185,22 @@ public class AdminService {
         }
     }
 
+    @Transactional
     public ResponseEntity<?> deleteUserConfirm(Long userid) throws IOException, InterruptedException {
+        Optional<User> userOptional = userRepository.findById(userid);
+        if (userOptional.isPresent()) {
+            User user = userOptional.get();
 
-        if (userRepository.findById(userid).isPresent()) {
+            User defaultDeletedUser = userRepository.findByUsername("deleted_user")
+                    .orElseGet(this::createDefaultDeletedUser);
 
-            User defaultDeletedUser;
-            Optional<User> defaultDeletedUserOptional = userRepository.findByUsername("deleted_user");
-            if (defaultDeletedUserOptional.isEmpty()) {
-                User defaultDeletedUserNew = new User();
-                defaultDeletedUserNew.setUsername("deleted_user");
-                defaultDeletedUserNew.setEmail("deleted_user@synergyaccounting.app");
-                defaultDeletedUserNew.setFirstName("deleted");
-                defaultDeletedUserNew.setLastName("user");
-                defaultDeletedUserNew.setUserType(UserType.ADMINISTRATOR);
-
-                UserDate userDate = new UserDate();
-                userDate.setBirthday(new Date());
-                userDate.setJoinDate(new Date());
-                userDate.setUser(defaultDeletedUserNew);
-
-                UserSecurity userSecurity = new UserSecurity();
-                String uuid = UUID.randomUUID().toString();
-                userSecurity.setPassword(passwordEncoder.encode(uuid));
-                userSecurity.setIsActive(false);
-                userSecurity.setIsVerified(true);
-                userSecurity.setFailedLoginAttempts(0);
-                userSecurity.setUser(defaultDeletedUserNew);
-
-                defaultDeletedUserNew.setUserSecurity(userSecurity);
-                defaultDeletedUserNew.setUserDate(userDate);
-
-                defaultDeletedUser = userRepository.save(defaultDeletedUserNew);
-
-            } else {
-                defaultDeletedUser = defaultDeletedUserOptional.get();
-            }
-
-            User user = userRepository.findById(userid).get();
-            UserSecurity userSecurity = user.getUserSecurity();
-
-            oldPasswordRepository.deleteAllByUserSecurity(userSecurity);
-            userSecurityRepository.delete(user);
-            userDateRepository.deleteByUser(user);
-
-            verificationRepository.deleteByUser(user);
-            tokenRepository.deleteByUser(user);
-            confirmationRepository.deleteByUser(user);
-
-            accountRepository.updateCreatorByCreator(user, defaultDeletedUser);
-            eventLogRepository.updateUserIdByUserId(user.getUserid(), defaultDeletedUser.getUserid());
-            journalEntryRepository.updateCreatorByCreator(user, defaultDeletedUser);
+            updateAndDeleteUserReferences(user, defaultDeletedUser);
 
             File file = new File("/home/sweappdomain/demobackend/uploads/" + user.getUserid() + ".jpg");
             if (scriptService.mailboxExists(user)) scriptService.deleteMailbox(user);
             if (file.exists()) file.delete();
 
             userRepository.delete(user);
-
             return ResponseEntity.ok(new MessageResponse("Successfully deleted user: " + user.getUsername()));
 
         } else {
@@ -254,6 +217,58 @@ public class AdminService {
         user.getUserSecurity().setIsActive(status);
         userRepository.save(user);
         return ResponseEntity.ok(new MessageResponse("User: "+ (status ? "activated" : "deactivated") + "successfully"));
+    }
+
+    private User createDefaultDeletedUser() {
+        User defaultDeletedUserNew = new User();
+        defaultDeletedUserNew.setUsername("deleted_user");
+        defaultDeletedUserNew.setEmail("deleted_user@synergyaccounting.app");
+        defaultDeletedUserNew.setFirstName("deleted");
+        defaultDeletedUserNew.setLastName("user");
+        defaultDeletedUserNew.setUserType(UserType.ADMINISTRATOR);
+
+        UserDate userDate = new UserDate();
+        userDate.setBirthday(new Date());
+        userDate.setJoinDate(new Date());
+        userDate.setUser(defaultDeletedUserNew);
+
+        UserSecurity userSecurity = new UserSecurity();
+        String uuid = UUID.randomUUID().toString();
+        userSecurity.setPassword(uuid, passwordEncoder);
+        userSecurity.setIsActive(false);
+        userSecurity.setIsVerified(true);
+        userSecurity.setFailedLoginAttempts(0);
+        userSecurity.setUser(defaultDeletedUserNew);
+
+        defaultDeletedUserNew.setUserSecurity(userSecurity);
+        defaultDeletedUserNew.setUserDate(userDate);
+
+        return userRepository.save(defaultDeletedUserNew);
+    }
+
+    private void updateAndDeleteUserReferences(User user, User defaultDeletedUser) {
+
+        accountRepository.updateCreatorByCreator(user, defaultDeletedUser);
+        em.flush();
+        eventLogRepository.updateUserIdByUserId(user.getUserid(), defaultDeletedUser.getUserid());
+        em.flush();
+        journalEntryRepository.updateCreatorByCreator(user, defaultDeletedUser);
+        em.flush();
+
+        oldPasswordRepository.deleteAllByUserSecurity(user.getUserSecurity());
+        em.flush();
+        userSecurityRepository.delete(user.getUserSecurity());
+        em.flush();
+        userDateRepository.delete(user.getUserDate());
+        em.flush();
+
+        verificationRepository.deleteAllByUser(user);
+        em.flush();
+        tokenRepository.deleteAllByUser(user);
+        em.flush();
+        confirmationRepository.deleteAllByUser(user);
+        em.flush();
+
     }
 
 }
